@@ -1,7 +1,7 @@
 package ch.phildev.springphawtrix.web.rest.dto;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
 
 import lombok.Data;
 
