@@ -1,29 +1,34 @@
 package ch.phildev.springphawtrix.app.clock;
 
+import ch.phildev.springphawtrix.app.domain.AppRegistration;
+import ch.phildev.springphawtrix.app.domain.PhawtrixApp;
+import ch.phildev.springphawtrix.app.management.AppDrawingComponentHolder;
+import ch.phildev.springphawtrix.domain.Coordinates;
+import ch.phildev.springphawtrix.domain.MatrixFrame;
+import ch.phildev.springphawtrix.service.MatrixTextCommandProvider;
+import ch.phildev.springphawtrix.web.rest.dto.DrawTextDto;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+import reactor.core.scheduler.Schedulers;
+
 import java.time.Duration;
 import java.time.Instant;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
 
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Component;
-import reactor.core.publisher.Flux;
-import reactor.core.scheduler.Schedulers;
-
-import ch.phildev.springphawtrix.app.domain.AppRegistration;
-import ch.phildev.springphawtrix.app.domain.PhawtrixApp;
-import ch.phildev.springphawtrix.domain.Coordinates;
-import ch.phildev.springphawtrix.domain.MatrixFrame;
-import ch.phildev.springphawtrix.service.MatrixTextCommandProvider;
-import ch.phildev.springphawtrix.web.rest.dto.DrawTextDto;
-
-@AllArgsConstructor
-@Component
 public class SimpleClock implements PhawtrixApp {
 
     private final AppRegistration appRegistration;
     private final MatrixTextCommandProvider matrixTextCommandProvider;
+    private Flux<Long> intervalFlux;
+
+    public SimpleClock(AppRegistration appRegistration,
+                       AppDrawingComponentHolder appDrawingComponentHolder) {
+
+        this.appRegistration = appRegistration;
+        this.matrixTextCommandProvider = appDrawingComponentHolder.getMatrixTextCommandProvider();
+    }
 
     @Override
     public AppRegistration getAppRegistration() {
@@ -44,18 +49,20 @@ public class SimpleClock implements PhawtrixApp {
         DrawTextDto.DrawTextDtoBuilder drawTextDtoTemplate = DrawTextDto.builder()
                 .coordinates(clockCoordinates)
                 .hexTextColor("#000000");
-        return Flux.interval(Duration.ofSeconds(1L), Schedulers.single())
+        this.intervalFlux = Flux.interval(Duration.ofSeconds(1L), Schedulers.single());
+        return intervalFlux
                 .map(interval -> MatrixFrame.builder()
                         .frameNumber(interval)
                         .frameBuffer(
                                 List.of(matrixTextCommandProvider.textDeliverFrame(
-                                        drawTextDtoTemplate.text(timeFormatter.format(Instant.now().truncatedTo(ChronoUnit.SECONDS))).build()
+                                        drawTextDtoTemplate.text(timeFormatter.format(Instant.now().truncatedTo(ChronoUnit.SECONDS)))
+                                                .build()
                                 ))
                         ).build());
     }
 
     @Override
-    public void stop() {
-
+    public Mono<Void> stop() {
+        return Mono.empty();
     }
 }
