@@ -1,17 +1,19 @@
 package ch.phildev.springphawtrix.app.management;
 
+import ch.phildev.springphawtrix.app.domain.AppRegistration;
+import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
+import org.jetbrains.annotations.NotNull;
+import org.springframework.util.Assert;
+import reactor.core.publisher.Mono;
+
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-import ch.phildev.springphawtrix.app.domain.AppRegistration;
-import lombok.Getter;
-import org.jetbrains.annotations.NotNull;
-import org.springframework.util.Assert;
-import reactor.core.publisher.Mono;
-
 @Getter
+@Slf4j
 public class DefaultInMemoryAppRegistrationRepository implements AppRegistrationRepository, Iterable<AppRegistration>{
 
     private final Map<String,AppRegistration> appRegistrations;
@@ -26,7 +28,8 @@ public class DefaultInMemoryAppRegistrationRepository implements AppRegistration
 
         appRegistrations.forEach(app -> {
             Assert.notNull(appRegistrations, "AppRegistration must not be null");
-            if (theMap.putIfAbsent(app.getAppName(), app) == null) {
+            if (theMap.putIfAbsent(app.getAppName(), app) != null) {
+                log.debug(app.getAppName());
                 throw new IllegalStateException(String.format("App already registered %s", app.getAppName()));
             }
         });
