@@ -1,21 +1,21 @@
 package ch.phildev.springphawtrix.app.management;
 
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
+import ch.phildev.springphawtrix.app.domain.AppRegistration;
+import ch.phildev.springphawtrix.app.management.config.PhawtrixAppProperties;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.springframework.boot.context.properties.PropertyMapper;
 
-import ch.phildev.springphawtrix.app.domain.AppRegistration;
-import ch.phildev.springphawtrix.app.management.config.PhawtrixAppProperties;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 @NoArgsConstructor(force = true,access = AccessLevel.PRIVATE)
 public final class PhawtrixAppPropertiesAppRegistrationAdapter {
 
     public static Map<String, AppRegistration> getAppRegistrations(PhawtrixAppProperties properties) {
-        return properties.getRegistration().entrySet().stream()
+        return properties.getRegistration()
+                .entrySet().stream()
                 .map(appEntry -> getAppRegistration(appEntry.getKey(), appEntry.getValue()))
                 .collect(Collectors.toMap(
                         AppRegistration::getAppName,
@@ -24,16 +24,16 @@ public final class PhawtrixAppPropertiesAppRegistrationAdapter {
     }
 
     private static AppRegistration getAppRegistration(String appId, PhawtrixAppProperties.Registration registration){
-        AppRegistration.AppRegistrationBuilder appBuilder = AppRegistration.builder().appName(appId);
+        AppRegistration.AppRegistrationBuilder appRegistrationBuilder = AppRegistration.builder().appName(appId);
 
         PropertyMapper  mapper = PropertyMapper.get().alwaysApplyingWhenNonNull();
-        mapper.from(registration::getAppId).to(appBuilder::appName);
-        mapper.from(registration::getAuthorName).to(appBuilder::appAuthor);
-        mapper.from(registration::getVersion).to(appBuilder::version);
+        mapper.from(registration::getAppId).to(appRegistrationBuilder::appName);
+        mapper.from(registration::getAuthorName).to(appRegistrationBuilder::appAuthor);
+        mapper.from(registration::getVersion).to(appRegistrationBuilder::version);
 
         //TODO: Arguments
 
-        return appBuilder.build();
+        return appRegistrationBuilder.build();
     }
 
 }
