@@ -1,21 +1,21 @@
 package ch.phildev.springphawtrix.web.rest;
 
+import ch.phildev.springphawtrix.config.WebSecurityConfig;
+import ch.phildev.springphawtrix.testutils.TestDataBuilderFactory;
+import ch.phildev.springphawtrix.web.rest.dto.AnswerDto;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
-import static org.hamcrest.Matchers.blankOrNullString;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.not;
-
-import ch.phildev.springphawtrix.testutils.TestDataBuilderFactory;
-import ch.phildev.springphawtrix.web.rest.dto.AnswerDto;
+import static org.hamcrest.Matchers.*;
 
 @WebFluxTest
 @Slf4j
+@Import(WebSecurityConfig.class)
 class DrawMatrixResourceTest {
 
     @Autowired
